# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controllers.ErrorResponse:
    properties:
      error: {}
    type: object
  controllers.ForgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
...
  controllers.ResetPasswordRequest:
    properties:
      new_password:
        type: string
      retype_password:
        type: string
      token:
        type: string
    type: object
  controllers.SimpleMessageResponse:
    properties:
      response: {}
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: API for brackets, matchups, voting, and engagement
  title: Matchup API
  version: "1.0"
paths:
...
  /password/forgot:
    post:
      consumes:
      - application/json
      description: Send a password reset email
      parameters:
      - description: Forgot password payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.SimpleMessageResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Request password reset
      tags:
      - auth
  /password/reset:
    post:
      consumes:
      - application/json
      description: Reset password using a valid token
      parameters:
      - description: Reset password payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.SimpleMessageResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Reset password
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Provide a valid JWT as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"